Email updates, subscriber newsletter form

+++ helpers/utils.py
@@ -6,6 +6,7 @@ from accounts.models import UserAffiliation
 from tklabels.models import TKLabel
 from bclabels.models import BCLabel
 from helpers.models import LabelTranslation, LabelVersion, LabelTranslationVersion
+from projects.models import ProjectActivity
 from xhtml2pdf import pisa
 
 from communities.models import Community, JoinRequest, InviteMember
@@ -14,6 +15,7 @@ from researchers.models import Researcher
 from .models import Notice
 from notifications.models import *
 
+from accounts.utils import get_users_name
 from helpers.emails import send_membership_email
 
 def check_member_role(user, organization):
@@ -156,28 +158,76 @@ def get_labels_json():
     data = json.load(json_data) #deserialize
     return data
 
-# Create/Update Notices
-def create_notices(selected_notices, organization, project, existing_notices):
+# Create/Update/Delete Notices
+def crud_notices(request, selected_notices, organization, project, existing_notices):
     # organization: either instance of institution or researcher
     # selected_notices would be a list: # attribution_incomplete # bcnotice # tknotice
-    if existing_notices:
-        for notice in existing_notices:
-            notice.delete()
+    # existing_notices: a queryset of notices that exist for this project already
+    name = get_users_name(request.user)
 
-    for selected in selected_notices:
-        notice_type = ''
-        if selected == 'bcnotice':
-            notice_type = 'biocultural'
-        elif selected == 'tknotice':
-            notice_type='traditional_knowledge'
-        elif selected == 'attribution_incomplete':
-            notice_type='attribution_incomplete'
-
+    def create(notice_type):
         if isinstance(organization, Institution):
-            Notice.objects.create(notice_type=notice_type, institution=organization, project=project)
+            new_notice = Notice.objects.create(notice_type=notice_type, institution=organization, project=project)
+            ProjectActivity.objects.create(project=project, activity=f'{new_notice.name} was applied to the Project by {name}')
 
         if isinstance(organization, Researcher):
-            Notice.objects.create(notice_type=notice_type, researcher=organization, project=project)
+            new_notice = Notice.objects.create(notice_type=notice_type, researcher=organization, project=project)
+            ProjectActivity.objects.create(project=project, activity=f'{new_notice.name} was applied to the Project by {name}')
+
+    def create_notices(existing_notice_types):          
+        for notice_type in selected_notices:
+            if notice_type:
+                if existing_notice_types:
+                    if not notice_type in existing_notice_types:  
+                        create(notice_type)
+                else:
+                    create(notice_type)
+    
+    if existing_notices:
+        existing_notice_types = []
+        for notice in existing_notices:
+            existing_notice_types.append(notice.notice_type)
+            if not notice.notice_type in selected_notices: # if existing notice not in selected notices, delete notice
+                notice.delete()
+                ProjectActivity.objects.create(project=project, activity=f'{notice.name} was removed from the Project by {name}')
+        create_notices(existing_notice_types)
+
+    else:
+        create_notices(None)
+
+def add_remove_labels(request, project, community):
+    # Get uuids of each label that was checked and add them to the project
+    bclabels_selected = request.POST.getlist('selected_bclabels')
+    tklabels_selected = request.POST.getlist('selected_tklabels')
+
+    bclabels = BCLabel.objects.filter(unique_id__in=bclabels_selected)
+    tklabels = TKLabel.objects.filter(unique_id__in=tklabels_selected)
+
+    user = get_users_name(request.user)
+
+    # find target community labels and clear those only!
+    if project.bc_labels.filter(community=community).exists():
+
+        for bclabel in project.bc_labels.filter(community=community).exclude(unique_id__in=bclabels_selected): # does project have labels from this community that aren't the selected ones?
+            project.bc_labels.remove(bclabel) 
+            ProjectActivity.objects.create(project=project, activity=f'{bclabel.name} Label was removed by {user} | {community.community_name}')
+
+    if project.tk_labels.filter(community=community).exists():
+        for tklabel in project.tk_labels.filter(community=community).exclude(unique_id__in=tklabels_selected):
+            project.tk_labels.remove(tklabel)
+            ProjectActivity.objects.create(project=project, activity=f'{tklabel.name} Label was removed by {user} | {community.community_name}')
+
+    for bclabel in bclabels:
+        if not bclabel in project.bc_labels.all(): # if label not in project labels, apply it
+            project.bc_labels.add(bclabel)
+            ProjectActivity.objects.create(project=project, activity=f'{bclabel.name} Label was applied by {user} | {community.community_name}')
+
+    for tklabel in tklabels:
+        if not tklabel in project.tk_labels.all():
+            project.tk_labels.add(tklabel)
+            ProjectActivity.objects.create(project=project, activity=f'{tklabel.name} Label was applied by {user} | {community.community_name}')
+    
+    project.save()
 
 
 def handle_label_versions(label):
